pyproject.toml
[project]
name = "crypto-dqn-ops"
version = "0.1.0"
description = "Cryptocurrency trading timing optimization using Rainbow DQN with MLOps best practices"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dvc>=3.65.0",
    "fastapi>=0.127.1",
    "fire>=0.7.1",
    "gymnasium>=1.1.1",
    "hydra-core>=1.3.2",
    "lightning>=2.6.0",
    "matplotlib>=3.9.4",
    "mlflow>=3.1.4",
    "numpy>=2.0.2",
    "onnx>=1.19.1",
    "onnxruntime>=1.20.1",
    "onnxscript>=0.5.7",
    "scipy>=1.13.1",
    "torch>=2.8.0",
    "uvicorn>=0.39.0",
]

[project.scripts]
crypto-dqn = "commands:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[dependency-groups]
dev = [
    "black>=25.11.0",
    "flake8>=7.3.0",
    "isort>=6.1.0",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
]
